requirements_cu11_6_torch_1_12_0.txt
# 基础工具包（整合去重+适配Python3.8+PyTorch1.13）
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==21.4b2
bleach==4.1.0
blessed==1.20.0
boto3==1.28.50
botocore==1.31.50
bottleneck==1.3.7
brotlipy==0.7.0
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cmake==4.2.3
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.7
cycler==0.11.0
Cython==3.0.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
docker-pycreds==0.4.0
easydict==1.13
einops==0.6.0  # 统一升级（兼容PyTorch1.13）
entrypoints==0.4
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.13.1
flit-core==3.9.0
fonttools==4.42.1
frozenlist==1.4.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221  # 无冲突，保留
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gorilla-core==0.2.7.8
gpustat==1.0.0
graphviz==0.8.4
grpcio==1.54.2
gtsam==4.2
huggingface-hub==0.17.1
hydra-core==1.1.1
idna==3.6
imageio==2.31.3
importlib-metadata==6.11.0
importlib-resources==6.0.1
iniconfig==2.0.0
iopath==0.1.10  # 与fvcore配套
ipdb==0.13.13
ipykernel==6.26.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==7.7.5
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jupyter-client==7.4.9
jupyter-core==5.5.0
jupyter-server==2.12.5
jupyterlab-pygments==0.3.0
jupyterlab-widgets==1.1.6
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
pytorch-lightning==1.9.1
lightning-utilities==0.9.0
lit==17.0.2
lmdb==1.4.1
loguru==0.7.2
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.7.5  # 适配numpy1.23.5+Python3.8
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mpmath==1.3.0
multidict==6.0.4
mxnet-mkl==1.6.0
mypy-extensions==1.0.0
mkl-service==2.4.0
mkl-fft==1.3.6
mkl-random==1.2.2
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==6.5.4
nbformat==5.5.0
nest-asyncio==1.5.8
networkx==3.1  # 统一升级（Python3.8兼容）
nltk==3.8.1
notebook==6.5.6
notebook-shim==0.2.3
numexpr==2.8.4
numpy==1.24.4  # 统一升级（Python3.8+PyTorch1.13最优版）
ninja==1.11.1
oauthlib==3.2.2
omegaconf==2.1.1
open3d==0.16.0  # 适配CUDA11.8
open-clip-torch==2.0.2
opencv-python-headless==4.8.0.76  # 替代opencv-python，避免冲突
packaging==23.2
pandas==1.5.3  # 降级（Python3.8兼容）
panopticapi @ https://github.com/cocodataset/panopticapi/archive/master.zip
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
pooch==1.8.0
portalocker==2.8.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3  # 核心降级（适配PyTorch3D0.7.2）
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pyDeprecate==0.3.1
pydot==1.4.2
pygments==2.17.2
pynvml==11.5.0
pyopenssl==23.3.0
pyparsing==3.1.1
pyre-extensions==0.0.23
pyrsistent==0.20.0
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
fastjsonschema==2.16.2
python-louvain==0.16
pytz==2023.3.post1
pyvista==0.42.2
PyWavelets==1.4.1
pyyaml==6.0.1
pyzmq==24.0.1
qudida==0.0.4
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
s3transfer==0.6.2
scikit-image==0.21.0
scikit-learn==1.2.2  
scipy==1.10.1  # 适配PyTorch1.13
scooby==0.7.4
send2trash==1.8.2
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
# stable-diffusion-sdkit==2.1.3
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorrt==8.5.3.1  # 适配CUDA11.8
termcolor==2.4.0
terminado==0.18.0
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.16  # 统一升级（适配PyTorch1.13）
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch-geometric==2.2.0  # 适配PyTorch1.13+cu117（单独安装）
torchmetrics==0.11.4
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
transformers==4.26.1
triton==2.0.0
trimesh==3.22.0
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.18
vtk==9.2.6
wandb==0.15.12
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.3.7
wheel==0.41.2
widgetsnbextension==3.6.6
xformers==0.0.16  # 适配PyTorch1.13.1+cu117
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.17.0

# 特殊安装项（需手动执行，不可放requirements.txt直接安装）
# 0 pytorch 1.12.0
# pip install torch==1.12.0+cu116 torchvision==0.13.0+cu116 torchaudio==0.12.0 --extra-index-url https://download.pytorch.org/whl/cu116
# 1. PyTorch3D（适配PyTorch1.13+cu117）
# pip install pytorch3d==0.7.2 -f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu117_pyt1130/download.html
# 2. PyG生态（适配PyTorch1.13+cu117）
# pip install torch-scatter torch-sparse torch-cluster torch-spline-conv -f https://data.pyg.org/whl/torch-1.12.0+cu116.html
# 3. detectron2（适配PyTorch1.13+cu117，替换原v0.6）
# pip install detectron2==0.6 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu117/torch1.13/index.html
# 4. lvis API
# pip install git+https://github.com/lvis-dataset/lvis-api.git@master
